# node-output/api/swagger.yaml
...
swagger: "2.0"
info:
  version: "v1"
  title: "Genius Integration Services API"
host: "eugen-dev.r02.xlgs.local"
basePath: "/XLCapital.XLI.GISiStage2/Gapi"
schemes:
- "https"
paths:
  /api/GetPoliciesByFilter:
    post:
      tags:
      - "DigitalPolicy"
      summary: "Get Policy Details when searching based on the Filter criteria"
      description: "If you enter policy search details like Policy Reference, inception\
        \ date expiry date,etc all the details related to the entered search criteria\
        \ will be displayed."
      operationId: "DigitalPolicy_Post"
      consumes:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "pageNumber"
        in: "query"
        description: ""
        required: true
        type: "string"
      - name: "pageSize"
        in: "query"
        description: ""
        required: true
        type: "string"
      - in: "body"
        name: "request"
        description: "The data in json format details the fields to create a new policy\
          \ event."
        required: true
        schema:
          $ref: "#/definitions/PolicyServiceFilterRequest"
      - name: "ProductId"
        in: "header"
        description: "Provide Product Id (Ex: NI01,DC01,GS01)"
        required: true
        type: "string"
      - name: "GlobalTrackingId"
        in: "header"
        description: "Provide Global Tracking Id"
        required: true
        type: "string"
      - name: "SourceSystem"
        in: "header"
        description: "Provide Source System Name"
        required: false
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/PolicyServiceFilterResponse"
      x-swagger-router-controller: "DigitalPolicy"
  /api/GetPolicyByReference:
    get:
      tags:
      - "DigitalPolicy"
      summary: "Get Policy Details when searching only by PolicyReference."
      description: "When you enter the policy reference, all the details related to\
        \ the entered policy reference will be displayed."
      operationId: "DigitalPolicy_Get"
      consumes: []
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "policyReference"
        in: "query"
        description: "The policy reference number e.g. SE00000847LI17A"
        required: true
        type: "string"
      - name: "ProductId"
        in: "header"
        description: "Provide Product Id (Ex: NI01,DC01,GS01)"
        required: true
        type: "string"
      - name: "GlobalTrackingId"
        in: "header"
        description: "Provide Global Tracking Id"
        required: true
        type: "string"
      - name: "SourceSystem"
        in: "header"
        description: "Provide Source System Name"
        required: false
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/PolicyServiceResponseExt"
      x-swagger-router-controller: "DigitalPolicy"
  /api/ConnectQuotesList:
    post:
      tags:
      - "Generic"
      summary: "Get a List of all the New/Renewal Quote numbers in Genius incepting\
        \ within the supplied date range."
      description: "If you enter inception start date, company code and line of business,\
        \ then a list of New/Renewal Quote numbers with related info will be retrieved\
        \ based on the provided search criteria."
      operationId: "Generic_Post"
      consumes:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - in: "body"
        name: "request"
        description: "The data in json format details the fields for Get Quotes List\
          \ Request."
        required: true
        schema:
          $ref: "#/definitions/ConnectQuotesListRequest"
      - name: "ProductId"
        in: "header"
        description: "Provide Product Id (Ex: NI01,DC01,GS01)"
        required: true
        type: "string"
      - name: "GlobalTrackingId"
        in: "header"
        description: "Provide Global Tracking Id"
        required: true
        type: "string"
      - name: "SourceSystem"
        in: "header"
        description: "Provide Source System Name"
        required: false
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/ConnectQuotesListResponse"
      x-swagger-router-controller: "Generic"
  /api/policies/{policyreferencenumber}/policyevents:
    post:
      tags:
      - "Policy"
      summary: "Add Policy Event entries to a Local Policy"
      description: "If you enter policy reference number along with policy event code,\
        \ event narrative and event effective date, then event gets added to the policy."
      operationId: "Policy_Post"
      consumes:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "policyreferencenumber"
        in: "path"
        description: "The policy reference number e.g. AT00015102PR18A"
        required: true
        type: "string"
      - in: "body"
        name: "request"
        description: "The data in json format details the fields to create a new policy\
          \ event."
        required: true
        schema:
          $ref: "#/definitions/AddPolicyEventRequest"
      - name: "globalTrackingId"
        in: "header"
        description: "Provide Global Tracking Id"
        required: true
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/AddPolicyEventResponse"
      x-swagger-router-controller: "Policy"
  /api/policies/{policyreferencenumber}/premiumstatus:
    get:
      tags:
      - "Policy"
      summary: "Get the summary Premium Collection Status of each Local Policy within\
        \ a Program."
      description: "If you enter the Master Policy Reference, then Premium Collection\
        \ Status of each Local Policy will be returned.\r\nIf you enter the Local\
        \ Policy Reference, then Premium Collection Status of that Local Policy will\
        \ be returned."
      operationId: "Policy_GetPremiumStatus"
      consumes: []
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "policyreferencenumber"
        in: "path"
        description: "The policy reference number e.g. US00085409PR18A"
        required: true
        type: "string"
      responses:
        "200":
          description: "OK"
          examples:
            application/json:
              $type: "Gis.Core.Api.Model.NI.GetProgramPremiumStatusResponse, Gis.Core.Api"
              PolicyPremiumStatuses:
              - LocalPolicyReference: "US00085409PR18A"
                PremiumStatus: "Full"
              MasterUnderwritingYear: 0
              MasterPolicyReference: "US00085409PR18A"
              RequestUser: "UserID"
              GeniusTransactionId: "0"
              Messages: []
              StatusMessages: []
              ProcessStatus: "Success"
          schema:
            $ref: "#/definitions/GetProgramPremiumStatusResponse"
      x-swagger-router-controller: "Policy"
  /api/policies/{policyreferencenumber}/ri:
    get:
      tags:
      - "Policy"
      summary: "Get a list of RI Attachments when searching by Policy."
      description: "When you enter the Policy the request will expect policy reference\
        \ and will bring back the RI attachments for the policy reference.\r\nThe\
        \ other fields are not in use for this version."
      operationId: "PolicyRI_Get"
      consumes: []
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "policyreferencenumber"
        in: "path"
        description: "The policy reference number e.g. US00085409PR18A"
        required: true
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/GetRIResponse"
      x-swagger-router-controller: "Policy"
    put:
      tags:
      - "Policy"
      summary: "Add RI Attachments details to a Local Policy"
      description: "If you enter policy reference number along with RI attachment\
        \ details, then RI details gets added to the policy."
      operationId: "PolicyRI_Put"
      consumes:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "policyreferencenumber"
        in: "path"
        description: "The policy reference number e.g. US00085409PR18A"
        required: true
        type: "string"
      - in: "body"
        name: "request"
        description: "The data in json format details the fields to create a new RI\
          \ details."
        required: true
        schema:
          $ref: "#/definitions/RIProcessingRequest"
      - name: "globalTrackingId"
        in: "header"
        description: "Provide Global Tracking Id"
        required: true
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/IGenericCommonResponse"
      x-swagger-router-controller: "Policy"
  /api/policies/{policyreferencenumber}/terms:
    put:
      tags:
      - "Policy"
      summary: "Update Policy terms field of a PolicyReference."
      description: "If you provide policy reference number along with payment condition\
        \ value,then it will be updated to the policy terms fields."
      operationId: "PolicyTerms_Put"
      consumes:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "policyreferencenumber"
        in: "path"
        description: "The policy reference number e.g. US00085409PR18A"
        required: true
        type: "string"
      - in: "body"
        name: "request"
        description: "The data in json format details for updating policy terms fields."
        required: true
        schema:
          $ref: "#/definitions/PolicyTermsRequest"
      - name: "globalTrackingId"
        in: "header"
        description: "Provide Global Tracking Id"
        required: true
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/IGenericCommonResponse"
      x-swagger-router-controller: "Policy"
  /api/policies/{policyreferencenumber}/transactions:
    get:
      tags:
      - "Policy"
      summary: "Get Policy transactions when searching by PolicyReference and type."
      operationId: "PolicyTransactions_Get"
      consumes: []
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "policyreferencenumber"
        in: "path"
        description: "The policy reference number e.g. DE00028392LI20A"
        required: true
        type: "string"
      - name: "type"
        in: "query"
        description: "Payment amount type"
        required: true
        type: "string"
      - name: "Authorization"
        in: "header"
        description: "Provide Bearer Token"
        required: true
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/PolicyTransactionsResponse"
      x-swagger-router-controller: "Policy"
  /api/programs:
    get:
      tags:
      - "Programs"
      summary: "Get a list of policies when searching by Program or Policy."
      description: "If you enter Program the request will expect master underwriting\
        \ year and program reference number.\r\nIf you enter Policy the request will\
        \ expect master policy reference number."
      operationId: "Programs_Get"
      consumes: []
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "programsearchmethod"
        in: "query"
        description: "Enter either Program or Policy."
        required: true
        type: "string"
      - name: "masterpolicyreferencenumber"
        in: "query"
        description: "Enter either Policy reference number or Quote reference number.(e.g.US00085409PR18A\
          \ or QUS0049669PR18A)"
        required: false
        type: "string"
      - name: "masterunderwritingyear"
        in: "query"
        description: "The year e.g. 2018"
        required: false
        type: "integer"
        format: "int32"
      - name: "programreferencenumber"
        in: "query"
        description: "The program e.g. 1199842"
        required: false
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/FindProgramResponse"
      x-swagger-router-controller: "Programs"
  /api/programs/{masterpolicyreferencenumber}:
    get:
      tags:
      - "Programs"
      summary: "Get Master and Local policies information."
      description: "If you enter master policy reference number, then Master and Local\
        \ policies information will be retrieved."
      operationId: "Programs_Get"
      consumes: []
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "masterpolicyreferencenumber"
        in: "path"
        description: "The master policy reference number e.g. SE00000847LI17A"
        required: true
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/GetProgramResponse"
      x-swagger-router-controller: "Programs"
...
definitions:
  PolicyServiceFilterRequest:
    type: "object"
    properties:
      PolicyReference:
        type: "array"
        items:
          type: "string"
      InceptionDate:
        type: "array"
        items:
          type: "string"
      ExpiryDate:
        type: "array"
        items:
          type: "string"
      UnderwritingYear:
        type: "array"
        items:
          type: "string"
      LineOfBusiness:
        type: "array"
        items:
          type: "string"
      BrokerId:
        type: "array"
        items:
          type: "string"
      InsuredId:
        type: "array"
        items:
          type: "string"
      CompanyId:
        type: "array"
        items:
          type: "string"
      BranchId:
        type: "array"
        items:
          type: "string"
      GlobalTrackingId:
        type: "string"
      SourceSystem:
        type: "string"
    example:
      PolicyReference:
      - ""
      - ""
      InceptionDate:
      - ">="
      - "20190601"
      ExpiryDate:
      - "<="
      - "20190830"
      UnderwritingYear:
      - ">"
      - "2019"
      LineOfBusiness:
      - "PROP"
      BrokerId:
      - "ACMI0009"
      InsuredId:
      - "AAAF0003"
      CompanyId:
      - "41"
      BranchId:
      - "E2"
  PolicyServiceFilterResponse:
    type: "object"
    properties:
      PolicyDetails:
        type: "array"
        items:
          $ref: "#/definitions/Policy"
      Reference:
        type: "string"
      GeniusTransactionId:
        type: "integer"
        format: "int32"
      Messages:
        type: "array"
        items:
          type: "string"
      StatusMessages:
        type: "array"
        items:
          $ref: "#/definitions/StatusMessage"
      ProcessStatus:
        type: "integer"
        format: "int32"
        enum:
        - 0
        - 1
        - 2
      GlobalTrackingId:
        type: "string"
      SourceSystem:
        type: "string"
    example:
      GeniusTransactionId: 6
      Messages:
      - "Messages"
      - "Messages"
      ProcessStatus: 1
      Reference: "Reference"
      GlobalTrackingId: "GlobalTrackingId"
      PolicyDetails:
      - MasterProductName: "MasterProductName"
        ExpiryDate: "2000-01-23T04:56:07.000+00:00"
        CountryOfOrigin: "CountryOfOrigin"
        InwardOutwardFlag: "InwardOutwardFlag"
        MainAccountCurrencyCode: "MainAccountCurrencyCode"
        NewRenewalFlag: "NewRenewalFlag"
        InceptionDate: "2000-01-23T04:56:07.000+00:00"
        MasterNumber: "MasterNumber"
        MasterProductSequence: "MasterProductSequence"
        MasterInternalStatus2: "Unknown"
        MasterStatusCode: "MasterStatusCode"
        MasterSequence: "MasterSequence"
        MasterTitle: "MasterTitle"
        SkeletonRiskFlag: "Skeleton"
        MasterProductCode: "MasterProductCode"
        MasterEndorsementNumber: "MasterEndorsementNumber"
        BranchNameCode: "BranchNameCode"
        BrokerNameCode: "BrokerNameCode"
        MasterReference: "MasterReference"
        CompanyNameCode: "CompanyNameCode"
        AmountsCurrencyCode: "AmountsCurrencyCode"
        InsuredNameCode: "InsuredNameCode"
        UnderwritingYear: 0
        AuthorizedFlag: "Authorized"
        MasterClass1Code: "MasterClass1Code"
      - MasterProductName: "MasterProductName"
        ExpiryDate: "2000-01-23T04:56:07.000+00:00"
        CountryOfOrigin: "CountryOfOrigin"
        InwardOutwardFlag: "InwardOutwardFlag"
        MainAccountCurrencyCode: "MainAccountCurrencyCode"
        NewRenewalFlag: "NewRenewalFlag"
        InceptionDate: "2000-01-23T04:56:07.000+00:00"
        MasterNumber: "MasterNumber"
        MasterProductSequence: "MasterProductSequence"
        MasterInternalStatus2: "Unknown"
        MasterStatusCode: "MasterStatusCode"
        MasterSequence: "MasterSequence"
        MasterTitle: "MasterTitle"
        SkeletonRiskFlag: "Skeleton"
        MasterProductCode: "MasterProductCode"
        MasterEndorsementNumber: "MasterEndorsementNumber"
        BranchNameCode: "BranchNameCode"
        BrokerNameCode: "BrokerNameCode"
        MasterReference: "MasterReference"
        CompanyNameCode: "CompanyNameCode"
        AmountsCurrencyCode: "AmountsCurrencyCode"
        InsuredNameCode: "InsuredNameCode"
        UnderwritingYear: 0
        AuthorizedFlag: "Authorized"
        MasterClass1Code: "MasterClass1Code"
      StatusMessages:
      - MessageCode: "MessageCode"
        MessageDescription: "MessageDescription"
        MessageDetail: "MessageDetail"
        MessageType: "Error"
      - MessageCode: "MessageCode"
        MessageDescription: "MessageDescription"
        MessageDetail: "MessageDetail"
        MessageType: "Error"
      SourceSystem: "SourceSystem"
...
  ConnectQuotesListRequest:
    type: "object"
    properties:
      IsMonthlyBatch:
        type: "boolean"
      InceptionStartDate:
        type: "string"
      CompanyCode:
        type: "array"
        items:
          type: "string"
      ProductCode:
        type: "array"
        items:
          type: "string"
      GlobalTrackingId:
        type: "string"
      SourceSystem:
        type: "string"
    example:
      IsMonthlyBatch: false
      InceptionStartDate: "20210101"
      CompanyCode:
      - "D7"
      - "B1"
      ProductCode:
      - "QLIABM"
      - "QEPLIE"
  ConnectQuotesListResponse:
    type: "object"
    properties:
      InceptionStartDate:
        type: "string"
      InceptionEndDate:
        type: "string"
      QuoteDetails:
        type: "array"
        items:
          $ref: "#/definitions/ConnectQuoteDetail"
      Reference:
        type: "string"
      GeniusTransactionId:
        type: "integer"
        format: "int32"
      Messages:
        type: "array"
        items:
          type: "string"
      StatusMessages:
        type: "array"
        items:
          $ref: "#/definitions/StatusMessage"
      ProcessStatus:
        type: "integer"
        format: "int32"
        enum:
        - 0
        - 1
        - 2
      GlobalTrackingId:
        type: "string"
      SourceSystem:
        type: "string"
    example:
      GeniusTransactionId: 6
      Messages:
      - "Messages"
      - "Messages"
      ProcessStatus: 1
      Reference: "Reference"
      GlobalTrackingId: "GlobalTrackingId"
      QuoteDetails:
      - ExpiryDate: "2000-01-23T04:56:07.000+00:00"
        ProductCode: "ProductCode"
        ProgramReference: "ProgramReference"
        DAMNameCode: "DAMNameCode"
        UnderwriterNameCode: "UnderwriterNameCode"
        PAMNameCode: "PAMNameCode"
        InsuredListName: "InsuredListName"
        QuoteReference: "QuoteReference"
        MainLineCode: "MainLineCode"
        CountryOfSettlementCode: "CountryOfSettlementCode"
        NewRenewalFlag: "NewRenewalFlag"
        InceptionDate: "2000-01-23T04:56:07.000+00:00"
        UnderwriterListName: "UnderwriterListName"
        PAMListName: "PAMListName"
        InsuredNameCode: "InsuredNameCode"
        UnderwritingYear: 0
        DAMListName: "DAMListName"
        CompanyCode: "CompanyCode"
        BranchCode: "BranchCode"
        QuoteStatusCode: "QuoteStatusCode"
        ProducingOffice: "ProducingOffice"
        QuoteTitle: "QuoteTitle"
      - ExpiryDate: "2000-01-23T04:56:07.000+00:00"
        ProductCode: "ProductCode"
        ProgramReference: "ProgramReference"
        DAMNameCode: "DAMNameCode"
        UnderwriterNameCode: "UnderwriterNameCode"
        PAMNameCode: "PAMNameCode"
        InsuredListName: "InsuredListName"
        QuoteReference: "QuoteReference"
        MainLineCode: "MainLineCode"
        CountryOfSettlementCode: "CountryOfSettlementCode"
        NewRenewalFlag: "NewRenewalFlag"
        InceptionDate: "2000-01-23T04:56:07.000+00:00"
        UnderwriterListName: "UnderwriterListName"
        PAMListName: "PAMListName"
        InsuredNameCode: "InsuredNameCode"
        UnderwritingYear: 0
        DAMListName: "DAMListName"
        CompanyCode: "CompanyCode"
        BranchCode: "BranchCode"
        QuoteStatusCode: "QuoteStatusCode"
        ProducingOffice: "ProducingOffice"
        QuoteTitle: "QuoteTitle"
      StatusMessages:
      - MessageCode: "MessageCode"
        MessageDescription: "MessageDescription"
        MessageDetail: "MessageDetail"
        MessageType: "Error"
      - MessageCode: "MessageCode"
        MessageDescription: "MessageDescription"
        MessageDetail: "MessageDetail"
        MessageType: "Error"
      InceptionEndDate: "InceptionEndDate"
      SourceSystem: "SourceSystem"
      InceptionStartDate: "InceptionStartDate"
...
  AddPolicyEventRequest:
    type: "object"
    properties:
      PolicyReference:
        type: "string"
      EventCode:
        type: "string"
      EventNarrative:
        type: "string"
      EventDate:
        type: "string"
        format: "date-time"
      RequestUser:
        type: "string"
      GlobalTrackingId:
        type: "string"
      SourceSystem:
        type: "string"
    example:
      PolicyReference: "AT00015102PR18A"
      EventCode: "BDS"
      EventNarrative: "Adding Event Narrative"
      EventDate: "2023-07-11T15:39:10.6745766+01:00"
      RequestUser: "UserID"
  AddPolicyEventResponse:
    type: "object"
    properties:
      PolicyReference:
        type: "string"
      RequestUser:
        type: "string"
      Reference:
        type: "string"
      GeniusTransactionId:
        type: "integer"
        format: "int32"
      Messages:
        type: "array"
        items:
          type: "string"
      StatusMessages:
        type: "array"
        items:
          $ref: "#/definitions/StatusMessage"
      ProcessStatus:
        type: "integer"
        format: "int32"
        enum:
        - 0
        - 1
        - 2
      GlobalTrackingId:
        type: "string"
      SourceSystem:
        type: "string"
    example:
      GeniusTransactionId: 0
      Messages:
      - "Messages"
      - "Messages"
      ProcessStatus: 6
      Reference: "Reference"
      GlobalTrackingId: "GlobalTrackingId"
      StatusMessages:
      - MessageCode: "MessageCode"
        MessageDescription: "MessageDescription"
        MessageDetail: "MessageDetail"
        MessageType: "Error"
      - MessageCode: "MessageCode"
        MessageDescription: "MessageDescription"
        MessageDetail: "MessageDetail"
        MessageType: "Error"
      PolicyReference: "PolicyReference"
      RequestUser: "RequestUser"
      SourceSystem: "SourceSystem"
  GetProgramPremiumStatusResponse:
    type: "object"
    properties:
      LocalPolicyReference:
        type: "string"
      PolicyPremiumStatuses:
        type: "array"
        items:
          $ref: "#/definitions/PolicyPremiumStatusData"
      ProgramReferenceNumber:
        type: "string"
      MasterUnderwritingYear:
        type: "integer"
        format: "int32"
      MasterPolicyReference:
        type: "string"
      RequestUser:
        type: "string"
      Reference:
        type: "string"
      GeniusTransactionId:
        type: "integer"
        format: "int32"
      Messages:
        type: "array"
        items:
          type: "string"
      StatusMessages:
        type: "array"
        items:
          $ref: "#/definitions/StatusMessage"
      ProcessStatus:
        type: "integer"
        format: "int32"
        enum:
        - 0
        - 1
        - 2
      GlobalTrackingId:
        type: "string"
      SourceSystem:
        type: "string"
...
  GetRIResponse:
    type: "object"
    properties:
      PolicyReference:
        type: "string"
      RIAttachments:
        type: "array"
        items:
          $ref: "#/definitions/RIAttachmentData"
      ProgramReferenceNumber:
        type: "string"
      MasterUnderwritingYear:
        type: "integer"
        format: "int32"
      MasterPolicyReference:
        type: "string"
      RequestUser:
        type: "string"
      Reference:
        type: "string"
      GeniusTransactionId:
        type: "integer"
        format: "int32"
      Messages:
        type: "array"
        items:
          type: "string"
      StatusMessages:
        type: "array"
        items:
          $ref: "#/definitions/StatusMessage"
      ProcessStatus:
        type: "integer"
        format: "int32"
        enum:
        - 0
        - 1
        - 2
      GlobalTrackingId:
        type: "string"
      SourceSystem:
        type: "string"
    example:
      GeniusTransactionId: 2
      Messages:
      - "Messages"
      - "Messages"
      RIAttachments:
      - CessionFlatAmount: 5.962133916683182
        TypeCode: "TypeCode"
        PolicyTitle: "PolicyTitle"
        ReinsurerGNSNameCode: "ReinsurerGNSNameCode"
        NetOfSequence: 6
        TypeDescription: "TypeDescription"
        FacTypeCode: "FacTypeCode"
        Sequence: 0
        ReinsurerName: "ReinsurerName"
        CessionPercentage: 1.4658129805029452
        PolicyReference: "PolicyReference"
        FacTypeDescription: "FacTypeDescription"
      - CessionFlatAmount: 5.962133916683182
        TypeCode: "TypeCode"
        PolicyTitle: "PolicyTitle"
        ReinsurerGNSNameCode: "ReinsurerGNSNameCode"
        NetOfSequence: 6
        TypeDescription: "TypeDescription"
        FacTypeCode: "FacTypeCode"
        Sequence: 0
        ReinsurerName: "ReinsurerName"
        CessionPercentage: 1.4658129805029452
        PolicyReference: "PolicyReference"
        FacTypeDescription: "FacTypeDescription"
      ProcessStatus: 7
      MasterUnderwritingYear: 5
      Reference: "Reference"
      StatusMessages:
      - MessageCode: "MessageCode"
        MessageDescription: "MessageDescription"
        MessageDetail: "MessageDetail"
        MessageType: "Error"
      - MessageCode: "MessageCode"
        MessageDescription: "MessageDescription"
        MessageDetail: "MessageDetail"
        MessageType: "Error"
      PolicyReference: "PolicyReference"
      RequestUser: "RequestUser"
      MasterPolicyReference: "MasterPolicyReference"
      ProgramReferenceNumber: "ProgramReferenceNumber"
      GlobalTrackingId: "GlobalTrackingId"
      SourceSystem: "SourceSystem"
...
  RIProcessingRequest:
    type: "object"
    properties:
      riDetails:
        type: "array"
        items:
          $ref: "#/definitions/RIAttachmentUpdateData"
      requestUser:
        type: "string"
      GlobalTrackingId:
        type: "string"
      SourceSystem:
        type: "string"
    example:
      riDetails:
      - calculationSequence: 40
        riPolicyNumber: "958846 031"
        netOfSequence: 0
        cessionPercentage: 5
        cessionFlatAmount: 0
        totalRICommission: ""
      requestUser: "UserID"
...
  IGenericCommonResponse:
    type: "object"
  PolicyTermsRequest:
    type: "object"
    properties:
      PaymentCondition:
        type: "string"
      requestUser:
        type: "string"
      GlobalTrackingId:
        type: "string"
      SourceSystem:
        type: "string"
    example:
      PaymentCondition: "P"
...